+++ servers/mcp-neo4j-data-modeling/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 ### Added
+* Add HTTP transport option
 
 ## v0.1.1